# Improve performance of in-sut diagnostic data reading