By only decoding the part of the in-situ data that contains the JSON numpy datatype the time spent in read_file() can be significantly reduced.

PR
```
CPU times: user 5.92 s, sys: 4.23 s, total: 10.1 s
Wall time: 11.6 s
```
Dev
```
CPU times: user 45.2 s, sys: 6.52 s, total: 51.7 s
Wall time: 53.2 
```

- [ ] **Small enough** (< few 100s of lines), otherwise it should probably be split into smaller PRs
- [ ] **Tested** (describe the tests in the PR description)
- [ ] **Runs on GPU** (basic: the code compiles and run well with the new module)
- [ ] **Contains an automated test** (checksum and/or comparison with theory)
- [ ] **Documented**: all elements (classes and their members, functions, namespaces, etc.) are documented
- [ ] **Constified** (All that can be `const` is `const`)
- [ ] **Code is clean** (no unwanted comments, )
- [ ] **Style and code conventions** are respected at the bottom of https://github.com/Hi-PACE/hipace
- [ ] **Proper label and GitHub project**, if applicable